tiny-embed-server/pyproject.toml
[project]
name = "tiny-embed-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.9.0",
    "fastapi>=0.116.1",
    "pillow>=11.3.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.54.0",
    "uvicorn>=0.35.0",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }
